new commit

+++ app/src/main/java/com/blogbasbas/mymovie/pojo/ResultsItem.java
@@ -35,7 +35,7 @@ public class ResultsItem{
 	private String releaseDate;
 
 	@SerializedName("vote_average")
-	private int voteAverage;
+	private double voteAverage;
 
 	@SerializedName("popularity")
 	private double popularity;
@@ -121,11 +121,11 @@ public class ResultsItem{
 		return releaseDate;
 	}
 
-	public void setVoteAverage(int voteAverage){
+	public void setVoteAverage(double voteAverage){
 		this.voteAverage = voteAverage;
 	}
 
-	public int getVoteAverage(){
+	public double getVoteAverage(){
 		return voteAverage;
 	}
